RestaurantsPage로 관심사 분리
- RestaurantsPage로 관심사 분리 및 테스트 코드 분리
- App 컴포넌트와 테스트 코드 수정
